Rename twigextensionmodule to twigextension
